move transactional invalidation tests to seperate `tests_transactions.py`
